+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 * Check if Vector Axis values is beetwen 2 others. ([6daf5eb](https://github.com/NazioLT/NZCore/commit/6daf5ebce600e69b847e963a7a09701a1e2e54ad))
 * Color set alpha method. ([f8d247b](https://github.com/NazioLT/NZCore/commit/f8d247b91c5566ee6a3cb4f4f7e5ec7c7c9974e3))
 * Bezier curves. ([de20f9e](https://github.com/NazioLT/NZCore/commit/de20f9e6da5267583d03c9d35c55a0132c302175))
+* Bezier Derivative. ([f60ab1d](https://github.com/NazioLT/NZCore/commit/f60ab1dc6894555bf036b584d3be0cc8d2a31ae6))
+* Bezier 2nd Derivative. ([]())
 
 #### Utils Library